erofs-utils: use virtual file interface for the buffer manager
In order to prepare for the upcoming metadata compression.

Link: https://lore.kernel.org/r/20250718065419.3338307-3-hsiangkao@linux.alibaba.com

+++ include/erofs/io.h
@@ -59,7 +59,7 @@ ssize_t erofs_io_pwrite(struct erofs_vfile *vf, const void *buf, u64 pos, size_t
 ssize_t erofs_io_pwritev(struct erofs_vfile *vf, const struct iovec *iov,
 			 int iovcnt, u64 pos);
 int erofs_io_fsync(struct erofs_vfile *vf);
-ssize_t erofs_io_fallocate(struct erofs_vfile *vf, u64 offset, size_t len, bool pad);
+int erofs_io_fallocate(struct erofs_vfile *vf, u64 offset, size_t len, bool pad);
 int erofs_io_ftruncate(struct erofs_vfile *vf, u64 length);
 ssize_t erofs_io_pread(struct erofs_vfile *vf, void *buf, u64 offset, size_t len);
 ssize_t erofs_io_read(struct erofs_vfile *vf, void *buf, size_t len);
